Refactor OpProcessor to enable simple overrides.
