--- gremlin-server/src/main/java/org/apache/tinkerpop/gremlin/server/op/standard/StandardOps.java
@@ -1,53 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.tinkerpop.gremlin.server.op.standard;
-
-import org.apache.tinkerpop.gremlin.driver.Tokens;
-import org.apache.tinkerpop.gremlin.driver.message.RequestMessage;
-import org.apache.tinkerpop.gremlin.server.Context;
-import org.apache.tinkerpop.gremlin.server.op.AbstractEvalOpProcessor;
-import org.apache.tinkerpop.gremlin.server.op.OpProcessorException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.script.Bindings;
-import javax.script.SimpleBindings;
-import java.util.Map;
-import java.util.Optional;
-
-/**
- * Operations to be used by the {@link StandardOpProcessor}.
- *
- * @author Stephen Mallette (http://stephen.genoprime.com)
- */
-final class StandardOps {
-    private static final Logger logger = LoggerFactory.getLogger(StandardOps.class);
-
-    public static void evalOp(final Context context) throws OpProcessorException {
-        final RequestMessage msg = context.getRequestMessage();
-
-        logger.debug("Sessionless request {} for eval in thread {}", msg.getRequestId(), Thread.currentThread().getName());
-
-        AbstractEvalOpProcessor.evalOp(context, context::getGremlinExecutor, () -> {
-            final Bindings bindings = new SimpleBindings();
-            Optional.ofNullable((Map<String, Object>) msg.getArgs().get(Tokens.ARGS_BINDINGS)).ifPresent(bindings::putAll);
-            return bindings;
-        });
-    }
-}
